+++ HISTORY.md
@@ -6,6 +6,11 @@ All notable changes to this project will be documented in this file.
 
 <!-- add unreleased items here -->
 
+* Fixed
+  * Properly closing output file (via [#]) 
+
+[#]: 
+
 ## 1.4.1 - 2026-06-30
 
 * Fixed